Tidy up metric model mappers - .map function for working with single objects - .mapper function for working with arrays
